Conflicts:
	openpdf/src/main/java/com/lowagie/text/Anchor.java
	openpdf/src/main/java/com/lowagie/text/Cell.java
	openpdf/src/main/java/com/lowagie/text/Chapter.java
	openpdf/src/main/java/com/lowagie/text/Chunk.java
	openpdf/src/main/java/com/lowagie/text/FontFactoryImp.java
	openpdf/src/main/java/com/lowagie/text/List.java
	openpdf/src/main/java/com/lowagie/text/MarkedSection.java
	openpdf/src/main/java/com/lowagie/text/Phrase.java
	openpdf/src/main/java/com/lowagie/text/Section.java
	openpdf/src/main/java/com/lowagie/text/SimpleCell.java
	openpdf/src/main/java/com/lowagie/text/SimpleTable.java
	openpdf/src/main/java/com/lowagie/text/VersionBean.java
	openpdf/src/main/java/com/lowagie/text/error_messages/MessageLocalization.java
	openpdf/src/main/java/com/lowagie/text/pdf/AcroFields.java
	openpdf/src/main/java/com/lowagie/text/pdf/BarcodeDatamatrix.java
	openpdf/src/main/java/com/lowagie/text/pdf/BaseField.java
	openpdf/src/main/java/com/lowagie/text/pdf/BaseFont.java
	openpdf/src/main/java/com/lowagie/text/pdf/ByteBuffer.java
	openpdf/src/main/java/com/lowagie/text/pdf/CFFFont.java
	openpdf/src/main/java/com/lowagie/text/pdf/CFFFontSubset.java
	openpdf/src/main/java/com/lowagie/text/pdf/CJKFont.java
	openpdf/src/main/java/com/lowagie/text/pdf/ColumnText.java
	openpdf/src/main/java/com/lowagie/text/pdf/FdfWriter.java
	openpdf/src/main/java/com/lowagie/text/pdf/FontDetails.java
	openpdf/src/main/java/com/lowagie/text/pdf/MappedRandomAccessFile.java
	openpdf/src/main/java/com/lowagie/text/pdf/PageResources.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfAcroForm.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfCell.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfChunk.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfContentByte.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfCopy.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfCopyFieldsImp.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfCopyFormsImp.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfDate.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfDocument.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfFormField.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfGraphics2D.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfLine.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfLister.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfNumberTree.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfPKCS7.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfReader.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfReaderInstance.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfSignatureAppearance.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfStamperImp.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfTable.java
	openpdf/src/main/java/com/lowagie/text/pdf/PdfWriter.java
	openpdf/src/main/java/com/lowagie/text/pdf/RandomAccessFileOrArray.java
	openpdf/src/main/java/com/lowagie/text/pdf/SequenceList.java
	openpdf/src/main/java/com/lowagie/text/pdf/SimpleBookmark.java
	openpdf/src/main/java/com/lowagie/text/pdf/SimpleNamedDestination.java
	openpdf/src/main/java/com/lowagie/text/pdf/TextField.java
	openpdf/src/main/java/com/lowagie/text/pdf/TrueTypeFont.java
	openpdf/src/main/java/com/lowagie/text/pdf/TrueTypeFontSubSet.java
	openpdf/src/main/java/com/lowagie/text/pdf/TrueTypeFontUnicode.java
	openpdf/src/main/java/com/lowagie/text/pdf/Type1Font.java
	openpdf/src/main/java/com/lowagie/text/pdf/Type3Font.java
	openpdf/src/main/java/com/lowagie/text/pdf/VerticalText.java
	openpdf/src/main/java/com/lowagie/text/pdf/XfaForm.java
	openpdf/src/main/java/com/lowagie/text/pdf/events/IndexEvents.java
	openpdf/src/main/java/com/lowagie/text/pdf/events/PdfPCellEventForwarder.java
	openpdf/src/main/java/com/lowagie/text/pdf/events/PdfPTableEventForwarder.java
	openpdf/src/main/java/com/lowagie/text/pdf/events/PdfPageEventForwarder.java
	openpdf/src/main/java/com/lowagie/text/pdf/fonts/cmaps/CMap.java
	openpdf/src/main/java/com/lowagie/text/pdf/fonts/cmaps/CMapParser.java
	openpdf/src/main/java/com/lowagie/text/pdf/hyphenation/SimplePatternParser.java
	openpdf/src/main/java/com/lowagie/text/pdf/internal/PdfAnnotationsImp.java
	openpdf/src/main/java/com/lowagie/text/pdf/internal/PdfViewerPreferencesImp.java
	openpdf/src/main/java/com/lowagie/text/pdf/parser/ParsedText.java
	openpdf/src/main/java/com/lowagie/text/pdf/parser/PdfContentReaderTool.java
	openpdf/src/main/java/com/lowagie/text/pdf/parser/PdfContentStreamHandler.java
	openpdf/src/main/java/com/lowagie/text/pdf/parser/PdfTextExtractor.java
	openpdf/src/main/java/com/lowagie/text/xml/XmlDomWriter.java
	openpdf/src/main/java/com/lowagie/text/xml/simpleparser/EntitiesToSymbol.java
	openpdf/src/main/java/com/lowagie/text/xml/simpleparser/EntitiesToUnicode.java
	openpdf/src/main/java/com/lowagie/text/xml/simpleparser/SimpleXMLParser.java
	openpdf/src/main/java/com/lowagie/text/xml/xmp/XmpArray.java
	openpdf/src/main/java/com/lowagie/text/xml/xmp/XmpReader.java
	openpdf/src/main/java/com/lowagie/text/xml/xmp/XmpWriter.java
	openpdf/src/test/java/com/lowagie/text/pdf/PdfSignatureAppearanceTest.java
	openpdf/src/test/java/com/lowagie/text/pdf/PdfTestBase.java
	openpdf/src/test/java/com/lowagie/text/pdf/SimplePdfTest.java
	openpdf/src/test/java/com/lowagie/text/pdf/TablePdfTest.java
	openpdf/src/test/java/com/lowagie/text/pdf/UnicodePdfTest.java
	pdf-html/src/main/java/com/lowagie/text/html/HtmlWriter.java
	pdf-html/src/main/java/com/lowagie/text/html/simpleparser/FactoryProperties.java
	pdf-html/src/main/java/com/lowagie/text/html/simpleparser/HTMLWorker.java
	pdf-html/src/main/java/com/lowagie/text/html/simpleparser/IncTable.java
	pdf-toolbox/src/main/java/com/lowagie/toolbox/plugins/Concat.java
	pdf-toolbox/src/main/java/com/lowagie/toolbox/plugins/ConcatN.java
	pdf-toolbox/src/test/java/com/lowagie/examples/directcontent/pageevents/Events.java
	pdf-toolbox/src/test/java/com/lowagie/examples/fonts/getting/RegisterFont.java
	pdf-toolbox/src/test/java/com/lowagie/examples/forms/FormList.java
	pdf-toolbox/src/test/java/com/lowagie/examples/forms/ListFields.java
	pdf-toolbox/src/test/java/com/lowagie/examples/objects/chunk/Glossary.java
	pdf-toolbox/src/test/java/com/lowagie/examples/objects/tables/alternatives/JTable2Pdf.java
	pdf-xml/src/main/java/com/lowagie/text/pdf/XfdfReader.java
	pdf-xml/src/main/java/com/lowagie/text/xml/SAXiTextHandler.java

+++ openpdf/src/main/java/com/lowagie/text/pdf/PdfPageLabels.java
@@ -308,7 +308,7 @@ public class PdfPageLabels {
             } else {
                 numberStyle = PdfPageLabels.EMPTY;
             }
-            formats[k] = new PdfPageLabelFormat(key.intValue() + 1, numberStyle, prefix, pagecount);
+            formats[k] = new PdfPageLabelFormat(key + 1, numberStyle, prefix, pagecount);
         }
         return formats;
     }
